refacor in rename dirs and moved files
